requirements.txt
# Core
requests
python-dotenv
pyyaml
beautifulsoup4

# LLM — uncomment the one you use, update llm_scorer.py accordingly
# groq              # free tier, good for getting started
# openai            # also works with Groq, Together, any OpenAI-compatible API
# anthropic

# Messaging platform SDK — uncomment the one you use, adapt notifier.py
# slack-sdk         # Slack
# slack-bolt        # Slack (interactive features: buttons, modals)
# discord.py        # Discord
# python-telegram-bot  # Telegram

# Search API — uncomment if using source_search.py
# google-search-results   # SerpAPI
# Any requests-based client works for Serper, Tavily, etc.

# Audio transcription — uncomment if using transcribe_call.py
# assemblyai
# openai            # for Whisper API

# Live recording — uncomment if using --record flag in transcribe_call.py
# sounddevice
# numpy
# scipy

# RSS source support — uncomment if adding RSS feeds as a source
# feedparser
